Move TickscriptSave button to own file
